requirements.txt
requests==2.27.1
pandas==1.4.1
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3